+++ tests/parsing.rs
@@ -16,8 +16,8 @@ fn test_package_segment_boundaries() -> Result<(), Box<dyn std::error::Error>> {
     assert_boundaries(common::pkgs::v4::RPM_BASIC_ECDSA_SIGNED.as_ref())?;
     assert_boundaries(common::pkgs::v4::RPM_BASIC_RSA_SIGNED.as_ref())?;
     assert_boundaries(common::pkgs::v4::RPM_BASIC_IMA_SIGNED.as_ref())?;
-    assert_boundaries(common::pkgs::v4::RPM_EMPTY_SRC.as_ref())?;
-    assert_boundaries(common::pkgs::v4::RPM_BASIC_SRC.as_ref())?;
+    assert_boundaries(common::pkgs::v4::src::RPM_EMPTY_SRC.as_ref())?;
+    assert_boundaries(common::pkgs::v4::src::RPM_BASIC_SRC.as_ref())?;
 
     let mut temp = tempfile::NamedTempFile::new()?;
 
@@ -31,8 +31,7 @@ fn test_package_segment_boundaries() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(feature = "signature-meta")]
     {
         use rpm::signature::pgp::Signer;
-        let signing_key = common::keys::v4::rsa_private();
-        let signer = Signer::load_from_asc_bytes(signing_key.as_ref())?;
+        let signer = Signer::load_from_asc_file(common::keys::v4::RSA_4K_PRIVATE)?;
         let constructed_pkg_with_sig =
             rpm::PackageBuilder::new("empty-package", "0", "MIT", "x86_64", "")
                 .build_and_sign(signer)?;
@@ -94,7 +93,7 @@ fn test_package_segment_boundaries() -> Result<(), Box<dyn std::error::Error>> {
 #[test]
 #[ignore]
 fn test_file_attrs() -> Result<(), Box<dyn std::error::Error>> {
-    let package = Package::open(common::pkgs::v4::RPM_FILE_ATTRS)?;
+    let package = Package::open(common::pkgs::v6::RPM_FILE_ATTRS)?;
     let metadata = &package.metadata;
 
     assert_eq!(metadata.is_source_package(), false);
@@ -1004,7 +1003,7 @@ fn test_empty_package() -> Result<(), Box<dyn std::error::Error>> {
 #[test]
 #[ignore]
 fn test_empty_source_package() -> Result<(), Box<dyn std::error::Error>> {
-    let package = Package::open(common::pkgs::v4::RPM_EMPTY_SRC)?;
+    let package = Package::open(common::pkgs::v4::src::RPM_EMPTY_SRC)?;
     let metadata = &package.metadata;
 
     assert_eq!(metadata.is_source_package(), true);